+++ login.php
@@ -45,9 +45,29 @@ $action = $_GET['action'] ?? '';
 					// Display the 'Forgot Password' form
 					$rs_login->forgotPasswordForm();
 					break;
+				case 'reset_password':
+					// Fetch all cookie keynames
+					$cookies = array_keys($_COOKIE);
+					
+					// Loop through the cookies
+					foreach($cookies as $cookie) {
+						// Check whether the reset password cookie is set
+						if(strpos($cookie, 'pw-reset') !== false) {
+							// Fetch the cookie's name
+							$cookie_name = $cookie;
+							break;
+						}
+					}
+					
+					// Fetch the cookie's data
+					$cookie_data = explode(':', $_COOKIE[$cookie_name]);
+					
+					// Display the 'Reset Password' form
+					$rs_login->resetPasswordForm($cookie_data);
+					break;
 				default:
-					// Display the login form
-					$rs_login->loginForm();
+					// Display the 'Log In' form
+					$rs_login->logInForm();
 			}
 			?>
 		</div>